Backend/requirements.txt
deepface==0.0.93
fastapi>=0.123.5
fpdf==1.7.2
google-generativeai>=0.8.5
google-search-results==2.4.2
gtts==2.5.4
imread-from-url==0.1.3
angchain==0.3.23
langchain-community==0.3.21
langchain-google-genai>=2.0.10
langchain-groq==0.2.0
lxml-html-clean==0.4.2
newspaper3k==0.2.8
onnxruntime>=1.23.2
openai==1.70.0
pillow>=12.0.0
pymongo==4.11.3
python-multipart>=0.0.20
pyttsx3==2.98
pyzbar==0.1.9
sentence-transformers==4.0.2
tf-keras==2.19.0
uvicorn>=0.38.0